+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Support on_*** observers on GWidget's directly
 - Support on_event_*** observers for SWT.constant event listeners (like show and hide)
 - Added widget focus listener and data-binding support
+- Support Glimmer DSL observe keyword and make it return observer registration object to unregister later on (unobserve)
+
 
 ## 0.4.8
 - Video widget